functions/postgres_to_raw/requirements.txt
# Azure Functions
azure-functions>=1.14.0

# Postgres and ADLS Gen2 (used by the timer trigger)
psycopg2-binary>=2.9.0
azure-identity>=1.15.0
azure-storage-file-datalake>=12.14.0
